basic info comp refactor
